requirements.txt
termcolor
PyGithub